backend/requirements.txt
flask
flask_sqlalchemy
flask_cors
pymysql
tushare